requirements.txt
asgiref==3.8.1
asttokens==3.0.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cs50==9.3.4
debugpy==1.8.12
decorator==5.1.1
dj-database-url==2.3.0
Django==5.1.6
django-storages==1.14.6
djangorestframework==3.15.2
executing==2.2.0
Flask==3.0.3
greenlet==3.1.1
gunicorn==23.0.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pywin32==308
pyzmq==26.2.1
six==1.17.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
termcolor==2.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
whitenoise==6.9.0
